Derive Macro communication::actor::firewall::RequestPermissions

#[derive(RequestPermissions)]

Implements the [VariantPermission] for struct/ unions with PermissionValue(1). For enums, it implements [ToPermissionVariants], which creates an according new enum Permission with Unit variants, and implements [VariantPermission] by assigning different [PermissionValue] for each variant. The permission value is the "index" in the enum as exponent for the power of 2, thus from top to bottom 1, 2, 4, 8...