Module communication::actor::types

Structs

EstablishedConnection
Information about the connection with a remote peer as maintained in the ConnectionManager.

Enums

CommunicationRequest
Requests for the [CommunicationActor].

CommunicationResults
Returned results from the [CommunicationActor]

ConnectPeerError
Errors that can occur in the context of a pending Connection.

FirewallBlocked
The firewall that rejected or dropped the request

KeepAlive
Determines if the local system should actively keep the connection alive

RelayConfig
Relay peer for outgoing request.

RequestMessageError